# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: License Key Delivery API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /keys:
    get:
      operationId: listKeys
      tags: [keys]
      summary: List license keys
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [available, assigned, delivered, failed]
      responses:
        "200":
          description: List of keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LicenseKey"
    post:
      operationId: createKey
      tags: [keys]
      summary: Create a single license key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LicenseKeyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseKey"
  /keys/import:
    post:
      operationId: importKeys
      tags: [keys]
      summary: Bulk import license keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyImportResult"
  /keys/{id}:
    delete:
      operationId: deleteKey
      tags: [keys]
      summary: Delete a license key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, fulfilled, failed, refunded]
        - name: source
          in: query
          schema:
            type: string
            enum: [manual, ebay, stripe]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Manually create and fulfill an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/fulfill:
    post:
      operationId: fulfillOrder
      tags: [orders]
      summary: Manually fulfill a pending/failed order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fulfilled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/resend-email:
    post:
      operationId: resendOrderEmail
      tags: [orders]
      summary: Re-send the license key email for a fulfilled order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendEmailResult"
        "404":
          description: Not found
        "400":
          description: Order not fulfilled

  # Stripe
...
components:
  schemas:
...
    LicenseKey:
      type: object
      required: [id, productId, keyValue, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        keyValue:
          type: string
        status:
          type: string
          enum: [available, assigned, delivered, failed]
        orderId:
          type: ["integer", "null"]
        assignedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        createdAt:
          type: string
    LicenseKeyInput:
      type: object
      required: [productId, keyValue]
      properties:
        productId:
          type: integer
        keyValue:
          type: string
          minLength: 1
    KeyImport:
      type: object
      required: [productId, keys]
      properties:
        productId:
          type: integer
        keys:
          type: array
          items:
            type: string
    KeyImportResult:
      type: object
      required: [imported, skipped, total]
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        total:
          type: integer
    Order:
      type: object
      required: [id, source, status, buyerEmail, productId, createdAt]
      properties:
        id:
          type: integer
        source:
          type: string
          enum: [manual, ebay, stripe]
        status:
          type: string
          enum: [pending, paid, fulfilled, failed, refunded]
        buyerEmail:
          type: string
        buyerName:
          type: ["string", "null"]
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        ebayOrderId:
          type: ["string", "null"]
        ebayLineItemId:
          type: ["string", "null"]
        stripePaymentIntentId:
          type: ["string", "null"]
        stripeSessionId:
          type: ["string", "null"]
        assignedKeyId:
          type: ["integer", "null"]
        assignedKeyValue:
          type: ["string", "null"]
        failureReason:
          type: ["string", "null"]
        fulfilledAt:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [buyerEmail, productId]
      properties:
        buyerEmail:
          type: string
        buyerName:
          type: string
        productId:
          type: integer
...
    ResendEmailResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        error:
          type: ["string", "null"]
...